Fix licensing terms and provide options for commercial use
The commit message indicates that the code changes involve fixing the licensing terms and providing options for commercial use. This commit updates the licensing terms to require for-profit companies and individuals intending to use the software for commercial use to obtain a commercial license. The commercial license allows proprietary use and source code modifications without the copyleft provisions of the GNU Affero General Public License version 3 (AGPLv3). The detailed description provides additional information about the commercial license and its purpose.